requirements.txt
pandas
streamlit
plotly
python-dotenv
SQLAlchemy
python-dateutil
